loading: define module before include, to solve this data-race
Passing the UUID via global state is a data-race, and passing it via TLS
seems more awkward than necessary. Instead, we define that `require`
implicitly makes the module it expects. We make the module a baremodule
initially, until the user calls `module Name`. That also sets the scope
for running the `__init__()` functions afterwards.